# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BrewDesk Café Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menu/categories:
    get:
      operationId: listCategories
      tags: [menu]
      summary: List all menu categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [menu]
      summary: Create a menu category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /menu/categories/{id}:
    patch:
      operationId: updateCategory
      tags: [menu]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [menu]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Menu Items ────────────────────────────────────────────────────────
  /menu/items:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List all menu items
      responses:
        "200":
          description: Menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu/items/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a menu item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    patch:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item (price, availability, discount)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      parameters:
        - name: date
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create a manual/quick-sale invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice (apply discount, change payment method)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  # ── Expenses ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, sortOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: ["string", "null"]
        sortOrder:
          type: integer
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        icon:
          type: string
        sortOrder:
          type: integer
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        sortOrder:
          type: integer
    MenuItem:
      type: object
      required: [id, name, price, categoryId, isAvailable]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        price:
          type: number
        discountPercent:
          type: ["number", "null"]
        finalPrice:
          type: number
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        preparationTime:
          type: ["integer", "null"]
    MenuItemInput:
      type: object
      required: [name, price, categoryId]
      properties:
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
        price:
          type: number
          minimum: 0
        discountPercent:
          type: number
          minimum: 0
          maximum: 100
        categoryId:
          type: integer
        isAvailable:
          type: boolean
        preparationTime:
          type: integer
    MenuItemUpdate:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
        price:
          type: number
          minimum: 0
        discountPercent:
          type: number
          minimum: 0
          maximum: 100
        categoryId:
          type: integer
        isAvailable:
          type: boolean
        preparationTime:
          type: integer
...
    InvoiceItem:
      type: object
      required: [description, quantity, unitPrice, subtotal]
      properties:
        description:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
    Invoice:
      type: object
      required: [id, sessionId, subtotal, discountAmount, tax, total, paymentMethod, status, createdAt, items]
      properties:
        id:
          type: integer
        sessionId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        subtotal:
          type: number
        discountAmount:
          type: number
        discountPercent:
          type: ["number", "null"]
        tax:
          type: number
        total:
          type: number
        amountPaid:
          type: ["number", "null"]
        change:
          type: ["number", "null"]
        paymentMethod:
          type: string
          enum: [cash, card, mixed]
        status:
          type: string
          enum: [open, paid, cancelled]
        staffId:
          type: ["integer", "null"]
        staffName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
    InvoiceInput:
      type: object
      required: [paymentMethod, items]
      properties:
        sessionId:
          type: integer
        customerName:
          type: string
        discountPercent:
          type: number
          minimum: 0
          maximum: 100
        amountPaid:
          type: number
          minimum: 0
        paymentMethod:
          type: string
          enum: [cash, card, mixed]
        staffId:
          type: integer
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
    InvoiceUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [open, paid, cancelled]
        discountPercent:
          type: number
        amountPaid:
          type: number
        paymentMethod:
          type: string
          enum: [cash, card, mixed]
        notes:
          type: string

    # ── Expenses ────────────────────────────────────────────────────────
...